requirements.txt
pandas
paho-mqtt
dash
plotly
redis